+++ src/screens/matching/MatchingScreen.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import { View, Text } from 'react-native';
 
-import Layout from '@/components/common/layout/Layout';
+import { Layout } from '@/components';
 
 export default function MatchingScreen() {
   return (
