+++ src/screens/onboarding/NotificationScreen.tsx
@@ -4,12 +4,7 @@ import React from 'react';
 import { useTranslation } from 'react-i18next';
 import { View } from 'react-native';
 import { useOnboardingStore } from '@/store/onboarding';
-import Button from '@/components/common/Button';
-import MyText from '@/components/common/MyText';
-import InnerLayout from '@/components/common/layout/InnerLayout';
-import Layout from '@/components/common/layout/Layout';
-import Heading from '@/components/onboarding/Heading';
-import HeadingDescription from '@/components/onboarding/HeadingDescription';
+import { Button, Heading, HeadingDescription, InnerLayout, Layout, MyText } from '@/components';
 
 export default function NotificationScreen({ navigation, route }) {
   const { t } = useTranslation('onboarding');
