+++ src/screens/onboarding/archived/StudentTypeSelectScreen.tsx
@@ -1,13 +1,9 @@
+import { Heading, InnerLayout, Layout, SelectItem } from '@/components';
 import React, { useState } from 'react';
 
 import { useTranslation } from 'react-i18next';
-import { Text, View } from 'react-native';
+import { Button, Text, View } from 'react-native';
 
-import Button from '@/components/common/Button';
-import InnerLayout from '@/components/common/layout/InnerLayout';
-import Layout from '@/components/common/layout/Layout';
-import Heading from '@/components/onboarding/Heading';
-import SelectItem from '@/components/onboarding/SelectItem';
 
 type StudentType = 'korean' | 'foreign' | null;
