+++ src/screens/verification/EmailCompleteScreen.tsx
@@ -1,13 +1,9 @@
+import { Button, Heading, HeadingDescription, Layout, MyText } from '@/components';
 import React from 'react';
 
 import { useTranslation } from 'react-i18next';
 import { View, Image } from 'react-native';
 
-import Button from '@/components/common/Button';
-import MyText from '@/components/common/MyText';
-import Layout from '@/components/common/layout/Layout';
-import Heading from '@/components/onboarding/Heading';
-import HeadingDescription from '@/components/onboarding/HeadingDescription';
 
 export default function EmailCompleteScreen({ navigation }) {
   const { t } = useTranslation('certification');
