+++ src/screens/verification/EmailScreen.tsx
@@ -1,20 +1,11 @@
+import { AuthRepository } from '@/api';
+import { ErrorMessage, FooterLayout, Heading, HeadingDescription, InnerLayout, KeyboardLayout, Label, Layout, MyText } from '@/components';
 import { Mail } from 'lucide-react-native';
 
 import React, { useState } from 'react';
 
 import { useTranslation } from 'react-i18next';
 import { View, TextInput } from 'react-native';
-
-import MyText from '@/components/common/MyText';
-import FooterLayout from '@/components/common/layout/FooterLayout';
-import InnerLayout from '@/components/common/layout/InnerLayout';
-import KeyboardLayout from '@/components/common/layout/KeyboardLayout';
-import Layout from '@/components/common/layout/Layout';
-import ErrorMessage from '@/components/onboarding/ErrorMessage';
-import Heading from '@/components/onboarding/Heading';
-import HeadingDescription from '@/components/onboarding/HeadingDescription';
-import Label from '@/components/onboarding/Label';
-import AuthRepository from '@/api/AuthRepository';
 
 const EMAIL_REGEX = /^[A-Za-z0-9]+$/;
