+++ src/screens/verification/StudentIdCompleteScreen.tsx
@@ -1,12 +1,9 @@
+import { Button, Heading, HeadingDescription, Layout } from '@/components';
 import React from 'react';
 
 import { useTranslation } from 'react-i18next';
 import { View } from 'react-native';
 
-import Button from '@/components/common/Button';
-import Layout from '@/components/common/layout/Layout';
-import Heading from '@/components/onboarding/Heading';
-import HeadingDescription from '@/components/onboarding/HeadingDescription';
 
 export default function StudentIdCardCompleteScreen({ navigation }) {
   const { t } = useTranslation('');
